+++ src/components/CopyToClipboard/index.ts
@@ -1,2 +1,2 @@
 export * from './CopyToClipboard';
-export type {CopyToClipboardStatus} from './types';
+export {CopyToClipboardStatus} from './types';
