+++ src/components/Label/Label.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import {block} from '../utils/cn';
-import {CopyToClipboard} from '../CopyToClipboard';
-import {CopyToClipboardStatus} from '../CopyToClipboard/types';
+import {CopyToClipboard, CopyToClipboardStatus} from '../CopyToClipboard';
 import {ClipboardIcon} from '../ClipboardIcon';
 import {Icon} from '../Icon';
 import {CrossIcon} from '../icons/CrossIcon';
