+++ src/components/ShareTooltip/ShareList/ShareList.tsx
@@ -7,8 +7,7 @@ import {ShareListItem} from '../ShareListItem/ShareListItem';
 import {Icon} from '../../Icon';
 import {Button} from '../../Button';
 import {Link} from '../../icons';
-import {CopyToClipboard} from '../../CopyToClipboard';
-import {CopyToClipboardStatus} from '../../CopyToClipboard/types';
+import {CopyToClipboard, CopyToClipboardStatus} from '../../CopyToClipboard';
 import {isOfType} from '../../utils/isOfType';
 
 import i18n from '../i18n';
